dev003 bugs

+++ core/include/update.inc.php
@@ -870,7 +870,9 @@ if ($this->getData(['core', 'dataVersion']) < 11600) {
 	$this->deleteData(['config','i18n', 'scriptGoogle']);
 	$this->deleteData(['config','i18n', 'showCredits']);
 	$this->deleteData(['config','i18n', 'autoDetect']);
-	$this->removeDir('core/vendor/i18n/css');
+	if (is_dir('core/vendor/i18n/css')) {
+		$this->removeDir('core/vendor/i18n/css');
+	}
 	// Nettoyage
 	if (file_exists('core/module/translate/ressource/googtrans.png')){
 		unlink('core/module/translate/ressource/googtrans.png');
@@ -878,8 +880,9 @@ if ($this->getData(['core', 'dataVersion']) < 11600) {
 	if (file_exists('core/vendor/i18n/inc.json')){
 		unlink('core/vendor/i18n/inc.json');
 	}
-	if (file_exits('core/vendor/i18n/translate.js')){
-		unlink('core/vendor/i18n/translate.js');}
+	if (file_exists('core/vendor/i18n/translate.js')){
+		unlink('core/vendor/i18n/translate.js');
+	}
 	if (file_exists('core/vendor/i18n/translation.js')){
 		unlink('core/vendor/i18n/translation.js');
 	}
